+++ learnCPP/learncpp.com/Chapter1/IO.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 
 // this function will fail with a ambiguous error in the compiler
-int cout()
+int16_t cout()
 {
 	return 5;
 }
@@ -18,7 +18,7 @@ int main() {
 	using namespace std;
 
 	cout << "Please enter any number!: ";
-	int myNumber{};
+	int16_t myNumber{};
 	cin >> myNumber;
 	cout << "You've entered: " << myNumber << endl;
 	return 0;
