+++ learnCPP/learncpp.com/Chapter1/forwardDeclartions.cpp
@@ -1,23 +1,23 @@
 #include <iostream>
 
-int additionNum(int firstNumber, int secondNumber, int thirdNumber);
+int16_t additionNum(int16_t firstNumber, int16_t secondNumber, int16_t thirdNumber);
 
 int main()
 {
 	std::cout << "Please insert the first number: ";
-	int firstNumber{};
+	int16_t firstNumber{};
 	std::cin >> firstNumber;
 	std::cout << "Please insert the second number: ";
-	int secondNumber{};
+	int16_t secondNumber{};
 	std::cin >> secondNumber;
 	std::cout << "Please insert the third number ";
-	int thirdNumber{};
+	int16_t thirdNumber{};
 	std::cin >> thirdNumber;
 	std::cout << "Total is: " << additionNum(firstNumber, secondNumber, thirdNumber) << std::endl;
 	return 0;
 }
 
-int additionNum(int firstNumber, int secondNumber, int thirdNumber)
+int16_t additionNum(int16_t firstNumber, int16_t secondNumber, int16_t thirdNumber)
 {
 	return firstNumber + secondNumber + thirdNumber;
 }
