+++ learnCPP/learncpp.com/Chapter1/functionCalls.cpp
@@ -3,14 +3,14 @@
 // return type of "void" means the function doesn't return a value so a return satement isn't needed
 void secondFunction() { std::cout << "We're in the secondFunction now!!!" << std::endl; }
 
-int myBirthday() { return 7; }
+int16_t myBirthday() { return 7; }
 
 void emptyFunction() { }
 
-int pickNumber()
+int16_t pickNumber()
 {
 	std::cout << "Pick a #: ";
-	int selectedNumber{};
+	int16_t selectedNumber{};
 	std::cin >> selectedNumber;
 	return selectedNumber;
 }
@@ -25,8 +25,8 @@ int main()
 	// below satement is invalid as you can't pass a void function into cout
 	// std::cout << emptyFunction() << " We got nothing back! emptyFunction using a void return type" << std::endl;
 
-	int x = pickNumber();
-	int y = pickNumber();
+	int16_t x = pickNumber();
+	int16_t y = pickNumber();
 
 	std::cout << "First # selected was: " << x << " followed by #: " << y << std::endl;
 	std::cout << "Total: " << x + y << std::endl;
