+++ learnCPP/learncpp.com/Chapter1/maths.h
@@ -1,7 +1,7 @@
 #ifndef MATHS_H
 #define MATHS_H
 
-int twoAddition(int firstNumber, int secondNumber);
-int twoMultiply(int firstNumber, int secondNumber);
+int16_t twoAddition(int16_t firstNumber, int16_t secondNumber);
+int16_t twoMultiply(int16_t firstNumber, int16_t secondNumber);
 
 #endif
