+++ learnCPP/learncpp.com/Chapter1/multipleFiles.cpp
@@ -10,17 +10,17 @@
 int main()
 {
 	std::cout << "#1: ";
-	int firstNum{};
+	int16_t firstNum{};
 	std::cin >> firstNum;
 	std::cout << "#2: ";
-	int secondNum{};
+	int16_t secondNum{};
 	std::cin >> secondNum;
 	std::cout << "Total after addition is: " << twoAddition(firstNum, secondNum) << std::endl;
 	std::cout << "#1: ";
-	int thirdNum{};
+	int16_t thirdNum{};
 	std::cin >> thirdNum;
 	std::cout << "#2: ";
-	int fourthNum{};
+	int16_t fourthNum{};
 	std::cin >> fourthNum;
 	std::cout << "Total after multiplication is: " << twoMultiply(thirdNum, fourthNum) << std::endl;
 	return 0;
