+++ learnCPP/learncpp.com/Chapter1/twoAddition.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-int twoAddition(int firstNumber, int secondNumber)
+int16_t twoAddition(int16_t firstNumber, int16_t secondNumber)
 {
 	return firstNumber + secondNumber;
 }
