+++ learnCPP/learncpp.com/Chapter1/twoMultiply.cpp
@@ -1,6 +1,6 @@
 #include <iostream>
 
-int twoMultiply(int firstNumber, int secondNumber)
+int16_t twoMultiply(int16_t firstNumber, int16_t secondNumber)
 {
 	return firstNumber * secondNumber;
 }
